refactor RESTful resources for mentorship applications and listings
